Remove AppDependency.swift
