+++ WordPress/Classes/Jetpack/JetpackAuthenticationManager.swift
