+++ WordPress/Classes/Jetpack/JetpackMigration/Common/MigrationAppDetection.swift
