+++ WordPress/Classes/Jetpack/JetpackMigration/Common/MigrationAppearance.swift
