+++ WordPress/Classes/Jetpack/JetpackMigration/Common/Views/Configuration/MigrationCenterViewConfiguration.swift
