+++ WordPress/Classes/Jetpack/JetpackMigration/Common/Views/Configuration/MigrationHeaderConfiguration.swift
