+++ WordPress/Classes/Jetpack/JetpackMigration/SuccessCard/CollectionView/DashboardMigrationSuccessCell.swift
@@ -21,6 +21,16 @@ class DashboardMigrationSuccessCell: UICollectionViewCell, Reusable {
         contentView.addSubview(view)
         contentView.pinSubviewToAllEdges(view)
     }
+
+    func configure(with viewController: UIViewController) {
+        self.onTap = { [weak viewController] in
+            guard let viewController else {
+                return
+            }
+            let handler = MigrationSuccessActionHandler()
+            handler.showDeleteWordPressOverlay(with: viewController)
+        }
+    }
 }
 
 extension DashboardMigrationSuccessCell: BlogDashboardCardConfigurable {
