+++ WordPress/Classes/Jetpack/JetpackMigration/SuccessCard/TableView/MigrationSuccessCell.swift
@@ -28,6 +28,17 @@ class MigrationSuccessCell: UITableViewCell {
     @objc func configureForSidebarMode() {
         cardView?.backgroundColor = .clear
     }
+
+    @objc(configureWithViewController:)
+    func configure(with viewController: UIViewController) {
+        self.onTap = { [weak viewController] in
+            guard let viewController else {
+                return
+            }
+            let handler = MigrationSuccessActionHandler()
+            handler.showDeleteWordPressOverlay(with: viewController)
+        }
+    }
 }
 
 extension BlogDetailsViewController {
