+++ WordPress/Classes/Jetpack/Utility/JetpackSocialError.swift
