+++ WordPress/Classes/Jetpack/Utility/SharedDataIssueSolver.swift
