+++ WordPress/Classes/System/JetpackWindowManager.swift
