--- WordPress/Jetpack/Classes/ViewRelated/WordPress-to-Jetpack Migration/Success card/Collection View/DashboardMigrationSuccessCell+Jetpack.swift
@@ -1,14 +1,0 @@
-import UIKit
-
-extension DashboardMigrationSuccessCell {
-
-    func configure(with viewController: UIViewController) {
-        self.onTap = { [weak viewController] in
-            guard let viewController else {
-                return
-            }
-            let handler = MigrationSuccessActionHandler()
-            handler.showDeleteWordPressOverlay(with: viewController)
-        }
-    }
-}
