--- WordPress/Jetpack/Classes/ViewRelated/WordPress-to-Jetpack Migration/Success card/Collection View/DashboardMigrationSuccessCell+WordPress.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-extension DashboardMigrationSuccessCell {
-
-    func configure(with viewController: UIViewController) {
-        // Adding an empty implementation of this method so Xcode doesn't complain.
-        // The whole `DashboardMigrationSuccessCell` shouldn't exist in WordPress. It's only needed for Jetpack.
-    }
-}
