--- WordPress/Jetpack/Classes/ViewRelated/WordPress-to-Jetpack Migration/Success card/Table View/MigrationSuccessCell+Jetpack.swift
@@ -1,15 +1,0 @@
-import Foundation
-
-extension MigrationSuccessCell {
-
-    @objc(configureWithViewController:)
-    func configure(with viewController: UIViewController) {
-        self.onTap = { [weak viewController] in
-            guard let viewController else {
-                return
-            }
-            let handler = MigrationSuccessActionHandler()
-            handler.showDeleteWordPressOverlay(with: viewController)
-        }
-    }
-}
