--- WordPress/Jetpack/Classes/ViewRelated/WordPress-to-Jetpack Migration/Success card/Table View/MigrationSuccessCell+WordPress.swift
@@ -1,10 +1,0 @@
-import Foundation
-
-extension MigrationSuccessCell {
-
-    @objc(configureWithViewController:)
-    func configure(with viewController: UIViewController) {
-        // Adding an empty implementation of this method so Xcode doesn't complain.
-        // The whole `MigrationSuccessCell` shouldn't exist in WordPress. It's only needed for Jetpack.
-    }
-}
